Split optional modules into their own file
