+++ ev3dev/brickpi.py
@@ -27,7 +27,7 @@ An assortment of classes modeling specific features of the BrickPi.
 """
 
 from .core import *
-
+from .optional import *
 
 OUTPUT_A = 'ttyAMA0:MA'
 OUTPUT_B = 'ttyAMA0:MB'
